mod_mantenimiento-mod_comunicacion/requirements.txt
flask
flask-sqlalchemy
flask_socketio
Flask-Login
Flask-JWT-Extended
Flask
werkzeug
reportlab
python-engineio
eventlet